split views of resource tracking in multiple file
