+++ resource_tracker/views/__init__.py
@@ -1,0 +1,4 @@
+from .resource_group_attribute_views import *
+from .resource_group_resource_views import *
+from .resource_group_views import *
+from .resource_pool_views import *
